@@ -2,8 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../../../core/models/game_event_model.dart';
 import '../../../core/models/game_model.dart';
+import '../../../core/models/player_model.dart';
 import '../../../core/services/firestore_service.dart';
-import 'manage_players_screen.dart'; // Reutiliza o playersStreamProvider
+import '../../players/screens/player_list_screen.dart';
 
 class ManageGameDetailsScreen extends ConsumerStatefulWidget {
   final Game game;
@@ -19,7 +20,6 @@ final gameEventsProvider = StreamProvider.autoDispose.family<List<GameEvent>, St
 
 class _ManageGameDetailsScreenState extends ConsumerState<ManageGameDetailsScreen> {
   
-  // --- NOVO MÉTODO ADICIONADO ---
   void _showEditScoreDialog(BuildContext context) {
     final ourScoreController = TextEditingController(text: widget.game.ourScore?.toString() ?? '');
     final opponentScoreController = TextEditingController(text: widget.game.opponentScore?.toString() ?? '');
@@ -53,12 +53,12 @@ class _ManageGameDetailsScreenState extends ConsumerState<ManageGameDetailsScree
 
                 if (ourScore != null && opponentScore != null) {
                   try {
-                    // Nota: O método updateGameScore precisa ser criado no FirestoreService
                     await ref.read(firestoreServiceProvider).updateGameScore(
                           widget.game.id,
                           ourScore,
                           opponentScore,
                         );
+                    // ignore: use_build_context_synchronously
                     if (mounted) Navigator.of(context).pop();
                   } catch (e) {
                     // TODO: Mostrar SnackBar de erro
@@ -73,13 +73,11 @@ class _ManageGameDetailsScreenState extends ConsumerState<ManageGameDetailsScree
     );
   }
 
-  void _showAddEventForm(BuildContext context) {
+  void _showAddEventForm(BuildContext context, List<Player> players) {
     final minuteController = TextEditingController();
     GameEventType selectedType = GameEventType.gol;
     String? selectedPlayerId;
     String? selectedPlayerName;
-
-    final playersAsyncValue = ref.read(playersStreamProvider);
 
     showDialog(
       context: context,
@@ -93,30 +91,30 @@ class _ManageGameDetailsScreenState extends ConsumerState<ManageGameDetailsScree
                   mainAxisSize: MainAxisSize.min,
                   children: [
                     DropdownButtonFormField<GameEventType>(
-                      value: selectedType,
+                      initialValue: selectedType,
                       items: GameEventType.values.map((type) => DropdownMenuItem(value: type, child: Text(_eventTypeToString(type)))).toList(),
                       onChanged: (value) => setStateDialog(() => selectedType = value!),
                       decoration: const InputDecoration(labelText: 'Tipo'),
                     ),
-                    playersAsyncValue.when(
-                      data: (players) => DropdownButtonFormField<String>(
-                        value: selectedPlayerId,
-                        hint: const Text('Selecione o Jogador'),
-                        items: players.map((p) => DropdownMenuItem(value: p.id, child: Text(p.name))).toList(),
-                        onChanged: (value) {
-                          setStateDialog(() {
-                            selectedPlayerId = value;
-                            selectedPlayerName = players.firstWhere((p) => p.id == value).name;
-                          });
-                        },
-                      ),
-                      loading: () => const Center(child: CircularProgressIndicator()),
-                      error: (e, s) => const Text('Erro ao carregar jogadores'),
+                    const SizedBox(height: 16),
+                    DropdownButtonFormField<String>(
+                      initialValue: selectedPlayerId,
+                      hint: const Text('Selecione o Jogador'),
+                      items: players.map((p) => DropdownMenuItem(value: p.id, child: Text(p.name))).toList(),
+                      onChanged: (value) {
+                        setStateDialog(() {
+                          selectedPlayerId = value;
+                          selectedPlayerName = players.firstWhere((p) => p.id == value).name;
+                        });
+                      },
+                      decoration: const InputDecoration(labelText: 'Jogador'),
+                      validator: (value) => value == null ? 'Campo obrigatório' : null,
                     ),
                     TextFormField(
                       controller: minuteController,
                       decoration: const InputDecoration(labelText: 'Minuto do Jogo'),
                       keyboardType: TextInputType.number,
+                      validator: (value) => value == null || value.isEmpty ? 'Campo obrigatório' : null,
                     ),
                   ],
                 ),
@@ -138,7 +136,7 @@ class _ManageGameDetailsScreenState extends ConsumerState<ManageGameDetailsScree
                         await ref.read(firestoreServiceProvider).addGameEvent(event);
                         if (mounted) Navigator.of(context).pop();
                       } catch (e) {
-                        // TODO: Mostrar SnackBar de erro
+                        // Mostrar um SnackBar de erro
                       }
                     }
                   },
@@ -164,7 +162,7 @@ class _ManageGameDetailsScreenState extends ConsumerState<ManageGameDetailsScree
             icon: const Icon(Icons.edit_note),
             tooltip: 'Editar Placar',
             onPressed: () {
-              _showEditScoreDialog(context); // --- MÉTODO CONECTADO AQUI ---
+              _showEditScoreDialog(context);
             },
           ),
         ],
@@ -219,10 +217,36 @@ class _ManageGameDetailsScreenState extends ConsumerState<ManageGameDetailsScree
           ),
         ],
       ),
+      // --- FLOATING ACTION BUTTON CORRIGIDO E ROBUSTO ---
       floatingActionButton: FloatingActionButton(
-        onPressed: () => _showAddEventForm(context),
-        child: const Icon(Icons.add),
+        onPressed: () async {
+          try {
+            ScaffoldMessenger.of(context).showSnackBar(
+              const SnackBar(content: Text('Carregando jogadores...'), duration: Duration(seconds: 1)),
+            );
+
+            final List<Player> players = await ref.read(allPlayersStreamProvider.future);
+            final playersInCategory = players.where((p) => p.categoryId == widget.game.categoryId).toList();
+
+            if (!mounted) return;
+
+            if (playersInCategory.isEmpty) {
+              ScaffoldMessenger.of(context).showSnackBar(
+                const SnackBar(content: Text('Não há jogadores cadastrados nesta categoria para adicionar eventos.')),
+              );
+            } else {
+              _showAddEventForm(context, playersInCategory);
+            }
+          } catch (e) {
+            if (mounted) {
+              ScaffoldMessenger.of(context).showSnackBar(
+                SnackBar(content: Text('Erro ao carregar jogadores: $e')),
+              );
+            }
+          }
+        },
         tooltip: 'Adicionar Evento',
+        child: const Icon(Icons.add),
       ),
     );
   }
