@@ -4,6 +4,7 @@ import 'package:go_router/go_router.dart';
 import 'package:intl/intl.dart';
 import '../../../core/models/news_model.dart';
 import '../../../core/services/firestore_service.dart';
+import '../../../shared/widgets/confirm_dialog.dart'; // 1. Importe o dialog
 
 final newsStreamProvider = StreamProvider.autoDispose<List<News>>((ref) {
   return ref.watch(firestoreServiceProvider).getNewsStream();
@@ -35,9 +36,16 @@ class ManageNewsScreen extends ConsumerWidget {
               subtitle: Text(DateFormat('dd/MM/yyyy').format(news.createdAt)),
               trailing: IconButton(
                 icon: const Icon(Icons.delete, color: Colors.red),
+                // 2. Modifique o onPressed para usar o dialog
                 onPressed: () async {
-                  // Adicionar confirmação antes de deletar
-                  await ref.read(firestoreServiceProvider).deleteNews(news.id);
+                  final confirm = await showConfirmDialog(
+                    context: context,
+                    title: 'Confirmar Exclusão',
+                    content: 'Tem certeza que deseja deletar a notícia "${news.title}"?',
+                  );
+                  if (confirm) {
+                    await ref.read(firestoreServiceProvider).deleteNews(news.id);
+                  }
                 },
               ),
             );
