@@ -1,29 +1,44 @@
 import 'package:flutter/material.dart';
-    import 'package:firebase_core/firebase_core.dart';
-    import 'package:flutter_riverpod/flutter_riverpod.dart';
-    import 'firebase_options.dart';
-    import 'navigation/app_router.dart';
-    import 'shared/theme/app_theme.dart'; // <-- 1. IMPORTE O TEMA
+import 'package:firebase_core/firebase_core.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
+import 'firebase_options.dart';
+import 'navigation/app_router.dart';
+import 'shared/theme/app_theme.dart';
 
-    void main() async {
-      WidgetsFlutterBinding.ensureInitialized();
-      await Firebase.initializeApp(
-        options: DefaultFirebaseOptions.currentPlatform,
-      );
-      runApp(const ProviderScope(child: MyApp()));
-    }
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp(
+    options: DefaultFirebaseOptions.currentPlatform,
+  );
+  runApp(const ProviderScope(child: MyApp()));
+}
 
-    class MyApp extends ConsumerWidget {
-      const MyApp({super.key});
+class MyApp extends ConsumerWidget {
+  const MyApp({super.key});
 
-      @override
-      Widget build(BuildContext context, WidgetRef ref) {
-        final router = ref.watch(goRouterProvider);
+  // A anotação @override deve estar exatamente aqui,
+  // logo acima do método que está sendo sobrescrito.
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final router = ref.watch(goRouterProvider);
 
-        return MaterialApp.router(
-          title: 'Moreira\'s Sports',
-          theme: appTheme, // <-- 2. APLIQUE O TEMA AQUI
-          routerConfig: router,
-        );
-      }
-    }
+    return MaterialApp.router(
+      title: 'Moreira\'s Sports',
+      theme: appTheme, // Usando a variável do seu arquivo de tema
+      
+      // Configuração para localização em Português (Brasil)
+      localizationsDelegates: const [
+        GlobalMaterialLocalizations.delegate,
+        GlobalWidgetsLocalizations.delegate,
+        GlobalCupertinoLocalizations.delegate,
+      ],
+      supportedLocales: const [
+        Locale('pt', 'BR'), // Adiciona suporte para pt_BR
+      ],
+      locale: const Locale('pt', 'BR'), // Define pt_BR como o padrão
+
+      routerConfig: router,
+    );
+  }
+}
